requirements.txt
# MetaTrader 5 Integration
MetaTrader5==5.0.5370

# AI & Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
pandas==2.2.0
numpy==1.26.3
ta==0.11.0
# ta-lib==0.4.28

# Technical Analysis
pandas-ta==0.3.14b0

# Telegram Bot
python-telegram-bot==20.7
aiohttp==3.9.3

# Database
redis==5.0.1
pymongo==4.6.1

# Async & Concurrency
asyncio
aiofiles==23.2.1

# Data Visualization
plotly==5.18.0
matplotlib==3.8.2

# Utilities
python-dotenv==1.0.1
pydantic
pyyaml==6.0.1
schedule==1.2.1
ccxt

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# VPS & Deployment
gunicorn==21.2.0
supervisor==4.2.5

# Windows Support (for MT5)
pytz==2024.1
